move conflict handling out of client into main
